+++ astroquery/esa/jwst/tests/DummyTapHandler.py
@@ -48,9 +48,7 @@ class DummyTapHandler:
         if method == self.__invokedMethod:
             return
         else:
-            raise ValueError(f"Method '+{str(method)}"
-                             + "' not invoked. (Invoked method is '"
-                             + f"{str(self.__invokedMethod)}"+"')")
+            raise ValueError(f"Method '+{str(method)}' not invoked. (Invoked method is '{str(self.__invokedMethod)}')")
 
     def check_parameters(self, parameters, method_name):
         print("FOUND")